Revert "More suspend timeout improvements"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f5751c76-19e4-4b29-bbbd-380fb01e42c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5737d94cdc48f2a5eb11a26a095c1c9aefc3d381
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com