Revert "Add stack type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f23bf32-0e79-4eef-9433-b131ce51e4d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic8c93a7dbabb41435395e2a28c1a0e821c1c5473
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com