Revert "Add stack type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f3b109ef-a3af-49b2-abc6-362c8eaba046). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7ed3bd18838a7807648e57b90a5a97342885806
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com