Revert "Add stack type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c785a431-d9a2-487e-8198-0fc0509e3536). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1b6210350d3c8c099516ef891d8fbd1e86c40cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com