Revert "Verify the presence of flag.info in ART APEXes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2eef1c84-4f6e-460e-a8ab-4b260443353e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iabe731d354a8c61c2d57dfc448f64edff7a86fd1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com