Revert "Add stack type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ffe6efe0-308d-444c-860b-a3b7a5f19f2c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5336f7bd75c82ffa117dd1bc100009f552c18a03
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com