Revert "Add stack type"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7f1dd85-f8da-4292-bdac-5bcde805ac5e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9e26f21973c0ad59f4586a466696b25d401709bc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com