Revert "Fix dex2oat CTS test to work in 64-bit only builds."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bc609b37-62b1-4afb-a984-2ccd38c4e2dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib3a7cc813b62325898f31c8177d7124cebf2f644
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com