Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4fab420-f750-4005-a730-56c7bd8d2c27). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6880df4e49c297bc8ec7eea6a63fe0641eb52955
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com