Consolidate various APEX properties to the ART shared defaults.
This doesn't change any of the APEXes except
test_broken_com.android.art, which now becomes a more complete APEX
with binaries, javalibs, and etc files. We rely on
unwanted_transitive_deps to avoid pulling in the non-broken libart.so
into it, so it no longer needs quite as much special treatment.

Test: art/build/apex/runtests.sh
Test: lunch armv8
      m test_broken_com.android.art test_jitzygote_com.android.art \
        test_imgdiag_com.android.art com.android.art \
        com.android.art.debug com.android.art.testing
  Check that the .apex files change in the expected way.
Bug: 370034097
Change-Id: I1ddf7f41bf5e6a4927e31ac842fc296acde7c910