Isolate a change that has a weird effect on the APEX builds.
Add some libraries to test_broken_com.android.art, to move it towards
being more complete like the other APEXes. Let's rely on
unwanted_transitive_deps to replace libart with libart-broken, instead
of leaving out all modules that in some way may depend on libart. This
is an intermediate change to pave the way for later CLs.

This should only change test_broken_com.android.art, but it turns out
it causes libjavacore to be built differently in _all_ the APEXes
defined in this file (and possibly elsewhere too).

Furthermore, it doesn't matter which APEXes contain
libcore_debug_native_shared_libs - if they're removed from exactly one
of the other APEXes then libjavacore is built bitwise identically
again. So it seems that libjavacore compilation output depends on the
_number_ of APEXes that have the libcore native libs, or specifically
either libandroidio or libopenjdk. That's even though the number of
APEXes that include libjavacore itself doesn't change.

This weirdness is separated out into a CL of its own since the child
CLs are tested by checking bitwise equality of the built APEXes.

Change-Id: I0e9636277b5dd22f1ead577490b03d41925b4889