Revert "Reland "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3bcc08f5-6ab4-4d7b-bd73-fd36079a5696). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61ae577eaae77723bca14fb98f43d5692201e1d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com