Revert "Reland "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9fd17425-fea4-43ad-af24-0e85f14ffc90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0cc16ae8a550ed3294f8b370760ced99480f714d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com