Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a73af5c-972d-49f3-9b50-d1b3af258d95). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0903753fd3d8bae3af137eb327db28fc4c38b4d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com