Revert "Reland "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3a64464-b788-464d-8ca3-5008180cf509). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I214065199f341433e37660a8cfe468effbeda6d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com