Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f5bfedd-120d-4249-8700-8cafca3aa28f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ad8482f741d84865213f3b4394c64a1744ded0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com