Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b0978fba-4238-4b1f-bf99-2a61140c17f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic0f255ef8e69572b7081d8c5b8991766c990bf87
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com