Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e4d75f5c-7542-4adc-a4c0-75fa72efe8db). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2602874566c0f74ebbe65bb2f48784f2f29264f8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com