Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a79dd21-6eb8-4213-ae32-2137c4fa9740). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6712f83d95592153349a62e0ae849eed65f0651f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com