Revert "Reduce memory used by `HEnvironment`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e49986b4-a144-4042-897c-6655e23cd0d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4fa36489561036c1d74064310f5e390fd6fd8da8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com