Revert "Reland "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/573c43ca-ec91-4d37-b2a4-a45ef3762d5e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I392a28489174fbfa222b6b517305f28970e09494
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com