Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36e839f2-1437-44ec-a8e4-d84ef3a1be90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fe897ef897b56c1ab85c7ab2acaed15b84777d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com