Don't bind-mount optional partitions if they don't exist.
On a device where an additional partition doesn't exist, the mount point
of the partition is a symlink to a directory inside /system. If we
bind-mount that at the corresponding symlink inside chroot, the mount
will follow the symlink and result at the location outside of chroot.

This CL also adds CHECKs to prevent similar things from happening.

Bug: 369736135
Test: atest CtsCompilationTestCases:android.compilation.cts.PreRebootDexoptTest
Change-Id: If9facfb2ebe07c5b9e7fc077a115c433ece7a325