Revert "Refactor the ART APEX Soong modules to make them easier ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46f404fc-4784-43c4-aea7-cff22c16b8ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic8bc12e9ecc729117e4fa9dd7308ec7ff3784cb0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com