Revert "Reduce memory used by `HEnvironment`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46f404fc-4784-43c4-aea7-cff22c16b8ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idc8c80e00215d931047126f6efd8441b6589e2dc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com