Revert "Reland "Calculate the number of out vregs.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8f391b1-0ef7-4eff-8bc3-d1dd4182c5f0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4d260741fe8878709c980e6be5d65f429a2adbc0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com