Revert "Reland "Calculate the number of out vregs.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05ba9de5-e4d1-4910-882d-743bba4e110c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I099f617bcacbb75200337c7f0e020e4b5cc75003
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com