Revert "Reland "Calculate the number of out vregs.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5ad74b6-4965-4067-a4f7-ee8787669a36). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id93659c61fb89c4b506c25b3eab17a3150c2ae71
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com