RFC: Avoid splitting 64KB mTHPs with kCollectorTypeCC
[Foreword: this RFC relates to a WIP effort for evaluating mTHP and
 contiguous PTEs on Android. This prototype CL is a first step to modify
 ART memory management and enable basic support for those features when
 running the concurrent-copying GC. As device support is still on-going,
 this CL has not yet been tested on an Android device running an
 mTHP-enabled kernel, and its impact on physical memory utilization,
 virtual memory fragmentation, and ART heap utilisation, is yet to be
 sufficiently studied. Once this is possible, follow-up work is intended
 to both optimize the implementation and support a wider range of ART
 memory management activities and configurations when running on a
 kernel with these features.]

With mTHP support, the kernel will greedily group mapped anonymous
memory into series of aligned multi-size transparent huge pages (mTHP,
also referred to as large folios), in powers-of-two sizes of the base
page size up to the PMD size. This optimizes memory management by
reducing page faults, as well as reducing kernel overhead due to fewer
objects for the kernel to maintain. On AArch64, this allows for an
additional optimization to be enabled for each aligned 64KB of
contiguous memory within an mTHP, where the contiguous PTE bit can be
used to allow a single TLB entry to map the entire 64KB contiguous
block, thereby reducing TLB pressure and improving performance.

Enabling both optimizations requires that the kernel can back the memory
into mTHPs that are multiples of 64KB, so it is important to not 'split'
them (e.g. into series of 4KB base pages) which would lose their
benefits as well as cause additional work to be necessary in order to
repopulate relevant kernel lists and TLB mappings. This means that
memory operations (e.g. madvise, mprotect, etc) should be done on 64KB
aligned boundaries wherever possible.

This prototype CL is a first step to supporting these features by
aligning ART heap memory management to 64KB regions, with the goal being
to maximize mTHP creation and avoid splitting 64KB mTHPs. As metrics
related to ART GC performance, heap fragmentation, heap utiliztation, and
so on, are yet to be studied on an Android device running an
mTHP-enabled kernel, support will be improved and optimized in
subsequent work.

For this prototype, a 4KB base page size and use of the CC collector is
currently required.

Notes on some of the relevant MemMaps:

* The main heap space (RegionSpace) used when running CollectorTypeCC is
  already managed in units of kRegionSize (256KB) and starts from a
  64KB-aligned base address, meaning that CC compaction of the main
  moving space already does not split 64KB contiguous blocks.
* The non-moving space is implemented by the external DlMalloc library
  which maps memory in units of the base page size. For this space to
  make optimal use of contiguous PTEs, the implementation should be
  changed to expand and release memory in blocks of 64KB. This is
  considered out of scope for this CL, but may be necessary for
  subsequent optimization efforts.
* LargeObjectSpace memory is allocated and freed in units of
  LargeObjectSpace::ObjectAlignment(). By increasing this value from the
  system page size (4KB) to 64KB, this CL ensures that the 64KB mTHPs
  allocated using contiguous PTEs during the initial LOS mmap, are not
  split during reclaimation. Note that aligning each large object to
  64KB will very likely result in a significant amount of wasted memory
  compared to 4KB and this will significantly impact heap utilisation.
  Subsequent optimization work should consider how objects are allocated
  to this space (e.g. increasing kMinLargeObjectThreshold or introducing
  other conditions).
* ELF images remain mapped according to kElfSegmentAlignment.
* AtomicStack mappings (mark stacks, object allocation stacks) and
  SpaceBitmaps (object live/marked bitmaps, object reference bitmaps)
  are mapped on 64KB aligned start addresses to maximize their presence
  in 64KB mTHP.

Change-Id: I7425ce3c192c9a5a69888d6f34f47a2b75f75a76