Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b70e718f-6eeb-411e-a919-ca2b749ab29a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3756ea4870792c3be250a702f51fc14c67b1dc2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com