Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07585b11-31a9-41ab-aa44-ed8742e78717). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2021d39a0dbaa7c834a59a5f515546f570e4a267
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com