Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/85982443-d124-41ed-9def-990c525ed9f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1808bbceef3a68589242c9c71a4847e6e0dd39cb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com