Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e9d73d19-cea1-4134-b297-cf6a91051a7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf56f165d4ec279a09be7607533fb37ab49c8e77
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com