IsMemorySharedMethod fix
It was checking the kAccMemorySharedMethod bit but that's
shared with the intrinsics so it can lead to the wrong result.

This is especially important for ArtMethod::UpdateCounter
since it skips updating the counter if IsMemorySharedMethod
is true. This meant we weren't updating the counter for
some intrinsics when we should have.

Test: art/test/testrunner/testrunner.py --host --64 -b --jit
Change-Id: I29cb1302d95edf0e328258f26b89429a513c9a22