Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18593dd0-577c-4c07-b099-592b9db2c5de). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6793973f7b4af9d2578076a05bf81b1db7bd3c2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com