Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12c26231-0a68-4fc3-91d2-65c70700bf8f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4fc00deddfc3dd608c6dc5284e20ae2fd0aaef48
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com