Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9504add8-e2c6-44d4-ab48-928bb33b4a2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89657e18694deade6bb7fbfda0783960a34a192e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com