Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b712103-b96a-4313-9afd-0bc3e5a6f441). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5825d550695d5fe229a658ff45054d545d21af78
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com