Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81b49e9b-9e8d-4cf5-a771-9d09a4bcc7e9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7bf3917a68b58e5ef48daf027dea090635f575ee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com