Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/504a151e-eec5-4255-89cf-f59f58012b07). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If46be931dbe243d61e9e2267fd4363710fce2902
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com