Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f88419e-93ca-4d88-846c-897397806f07). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I31a7e89deb1e2dd41a98d4ab1ee281a4697528b1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com