Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dcb95ddd-5114-4da7-ba58-37afa99ae0cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7224919489c193ad2919d80d030534357a94cc5f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com