Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b63805d1-6830-480f-b490-701c2cf9022f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5cfff8e630fe71b21bf7ed9c740aebb4858bbd36
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com