Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1b0299b2-21e9-45bd-b2f6-acbb3d975ec2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I23b3d753a591b38e25c45d80fe49b0627bc53faf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com