Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bf065dd0-d24c-428a-b128-e0f2d0f69871). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0754f5a6aaf00ad0d80066b688276ccf3bc6426d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com