Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/026a9bdf-ef8c-4961-8b6a-50042eee5e58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ba4906c2837f9ab9d1cf96a3e6cf79785d84745
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com