Revert "Revert "Add intrinsics for the absolute forms of unsafe...."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05ab266c-7ce4-4583-a915-fc8e75f360a5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifb689202d24640c283ff47da11b3e67fcc1febae
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com