Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fbf7b6c0-f6e2-4fe6-86c9-a3048e3b2f08). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If6f485ee0c27af6e611a0108f2d47ddd87549744
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com