Revert "IsMemorySharedMethod fix for intrinsics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a5991fb8-1c5d-4e0f-a123-4361fbbb5962). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373867801
Change-Id: I6c61619906fb085dd9ee7735c18a4834fe20a41f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com