Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36722ee9-7c53-40e9-a00b-2e8597c82fdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa0b07df10ba1550c99aa5a80dc15e5d646bed7e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com