Fix `--jit-on-first-use` run-tests.
For 570-checker-osr-locals, ensure that the `WeirdLoop`
class is visibly initialized for non-x86 architectures, so
that JIT-on-first-use enters the method in compiled code.
Since JIT-on-first-use never performs OSR compilation, the
method previously started executed in interpreter and it was
then waiting for OSR until the test timed out.

To fix 848-pattern-match, return from `ForceJitCompiled()`
without checking the entrypoint for JIT-on-first-use. When
the pattern matching succeeds, the entrypoint check would
wait forever.

Test: testrunner.py --target --jit-on-first-use
Change-Id: Iff09f067e36c4b1d8b7500c2b09b29b5c6b5ebd9