Revert "IsMemorySharedMethod fix for intrinsics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a5991fb8-1c5d-4e0f-a123-4361fbbb5962). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373867801
Change-Id: I2264a1ce8b745ea9b58e606daa82ee2f071b3d3f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com