Revert "IsMemorySharedMethod fix for intrinsics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7bc7250f-0a64-4d52-a135-64db42cb5cba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373873275
Change-Id: Ia03c8869ce91bfe4913902bda8a4d5c110d013c9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com