Update method tracing to work when always_enable_profile_code
Enabling always_enable_profile_code adds code to record entries into a
circular buffer at the method entry and exit. This code is gated by the
condition that the buffer is not a nullptr. We use the same entry to
hold buffer for both regular method tracing and low-overhead tracing.
To correctly support method tracing with always_enable_profile_code we
replace the entry points that record low overhead traces with a dummy
nop entry points that just return.

Bug: 352518093
Test: manual testing with feature enabled on 2246* tests
Change-Id: Ide39f54b8005d227db0a20e0d3e5634809d57ff5