Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ec1f908d-9dd2-442b-96fe-09913c02abc4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I539dd34f9904b53c36d1d54f4abd8b0214d07379
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com