Revert "Reland^2 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/80ef2161-4fca-4da1-a6ff-aacf549494df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I51149be48ab98df9a8591b3a2b0373a6254b9c8b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com