Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/db7f5dac-ea71-4c6a-85fe-af9fb5708f24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72187c3521ceabc17277bb651c331fb2577bb47f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com