Revert "Refactor `HandleCache` out of `nodes.{h,cc}`."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72bfca72-e2b9-461f-8ff5-70acf9136404). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5dbcfc4c0b73f308a703486ac270f9d8af0ef795
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com