Revert "Refactor run-test to use python argument parser."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89f4e09b-0795-4ceb-a0f2-b0b27f1db9aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5c3f4f68aa1b4d5b6e1c6a78b5cf4e44f42e4629
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com