Revert "Add `art_boot_images` as required dep of some ART tests"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/62daa366-9ae4-4f91-b65e-6c82abe9fad0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d8ce85bd99036297019fccd1ba4fd3a1ac13be8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com