Fix app info code type.
The app info class was introduced by https://r.android.com/1749100, and it
tags split apks as the base apk from day one. The reason behind this is
unclear. This may confuse other code that tries to identify the base apk.

This CL attempts to fix this, but it's admittedly a risky change. We will
closely monitor continuous testing and metrics from dogfood after
submitting it.

Change-Id: I3e23e99a317c9d389fe919bbeb2ea5565b92fe20
Test: Presubmit