Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0cdc3ee5-2ba7-4ca6-b48f-d2bc33bb054d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie229b96ba4a2caee971b6c15ea84b6d0541309d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com