Revert "Refactor run-test to use python argument parser."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/35578952-7aab-4a76-a7bd-4c37c98d358d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I546c754b5dab1d297b9b320afd452020c592f9f7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com