Revert "Move code around to compile when asan is enabled."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b95887ee-5061-45cb-b3bc-588cadaf1cab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id63234bb8cbfd3955905e6e99fc2f69748f51a2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com