Revert "Delete compact_dex_level.h."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb8777cc-b460-42f5-8b43-9ae2ea600cc9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ied4c7dd6bd48d289876e211448f293cb491d627d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com