Flush profiles upon `pm dump-profiles`.
Before this change, `pm dump-profiles` only dumps the profiles that are
already on disk. Users can force the app to flush the profiles from the
memory to the disk by sending `SIGUSR1` to app processes, but this is
not ideal:
1. The shell user cannot send `SIGUSR1`, meaning it doesn't work on user
   builds.
2. The user doesn't know when the profile save is done, so they need to
   conservatively wait a long time after sending `SIGUSR1`, typically 1
   second or more.

After this change, `pm dump-profiles` forces the app to flush the
profiles from the memory to the disk before dumping the profiles.

The profile flush is implemented in the following flow:
1. `artd` places notification files at
   `/data/misc/profiles/cur/<user_id>/<package_name>/pid_<pid>.tmp
2. For each app process:
   (1). `system_server` sends `SIGUSR1` to the app process.
   (2). The app process flushes profiles and deletes the notification
        file corresponding to its own pid.
   (3). `artd` waits for the nofitication file to be deleted.

Bug: 254102312
Test: atest ArtServiceTests
Test: atest art_standalone_artd_tests
Test: adb shell pm dump-profiles com.google.android.gms
Change-Id: I7ea2403f2ba6aa1fcbefa0f42a9ca51487aa7ef5