Revert "Split dirty-image-objects file between art and framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a7d6a32c-b25f-4caa-a49b-7e00e74c8db3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifa0b1cef93c7e99f8565acb2e4c85265fc055489
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com