Revert "Revert "Reland^2 "Introduce black-dense region in moving..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f58e589-3448-4d1a-b793-640ed97edf80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If10d088d74eac205b370f03ba3042a6fa3db0964
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com