Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/20f9facd-a2bc-40fc-8bbd-0fe3ec67751f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73deb22c63120e5f425665a63876ed9192e5d31a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com