Revert "run-test: Replace "true"/"false" with True/False"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c47061ef-3712-4d20-b002-f79eec59e57a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iff50eb0cb761fc98bdab7768382913a872676f5e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com