Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a1c2ee83-caf9-44c7-afe9-5cd2a2cd691a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idcc06f3728a91ecae167d320b95733ee50042cca
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com