Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/daf0d83b-a01e-4b2e-8c76-07515aea993a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf2518b8274e9fe22d741972b75eaea05776e3f2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com