Revert "Do not defer library unloading from zygote"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c7a63b5-be8c-4f51-b1f5-2f5a1f568222). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0191b3f7ebc4342f936c78dbbd8c5dee267fb797
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com