Revert "Reland^3 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0c7a63b5-be8c-4f51-b1f5-2f5a1f568222). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iece2a4923f58797111c79f6b52e97807b887b4c1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com