Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e4fe27e-a0e6-487d-93ab-42ce79b13573). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ieec302dcbfe8d53955c5375131e80a073e9037d3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com