Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1258d58a-c0c9-419d-87dd-69f19bf14551). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idd7aa25e709134360eafd74cb94cbf55aff83363
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com