Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71873f89-9751-461a-bef9-41cc20070ed2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c31089338a07534d73f470b7976812cfc5a682f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com