Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5863439d-4e4f-4d49-a3ad-69888c2fc6ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I91bf34241db6e67a3b23788f9097a2bb60716a3d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com