Revert "Add new HeapByteBuffer methods into profile"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7b2f6a3b-6877-4827-9137-18799f1603dd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia05f6c6f07f0bcd1b7be9840fd5d7c136f44109e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com