cleanup FixUpInstructionType
It was expecting to have an HSelect as an input, so we might
as well encode that with the C++ types. Rename it to
FixUpSelectType.

Also move the instructions around so that the ScopedObjectAccess
scope is smaller. SetReferenceTypeInfo acquires ScopedObjectAccess
in debug builds so we can avoid the nesting of ScopedObjectAccess
calls.

As a drive-by, move ScopedObjectAccess to CheckAgainstUpperBound
so that we only call it in a subset of cases.

Test: art/test/testrunner/testrunner.py --host --64 -b --optimizing
Change-Id: Id6f179e68e0a460577d5e42b8c431f3d035405a4