Update the doc of Thread::StartAssertNoThreadSuspension(cause)
StartAssertNoThreadSuspension(cause) doesn't seem to cause an
assertion failure when Thread::AllowThreadSuspension() is called,
and handles pending suspension requests, when the thread
could be on a runnable state. Update the doc to reflect it.

Context:
I am looking at where the switch interprepter runs checkpoint
functions. For example, the interprepter may run into a checkpoint
functions in thread->AllowThreadSuspension(), but the assertion doesn't
seem to cause an assertion failure. If I try to check
DCHECK(tls32_.no_thread_suspension == 0) in the Thread::CheckSuspend,
it will cause a native crash.

Test: doc update only
Change-Id: Ib44a0f319d13f50d768bda4087a2c3816a4a4cb9