Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/532c146c-43f9-41d5-9a7d-9e003ee6ef58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9414a53c3ec8d79cbed9d170e29a3d4239f702fc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com