Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72da37b1-fa89-44de-b8cd-5f6a429347aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I75f13b8da14a6f0a392db9243b0b4d482e2eebef
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com