Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72da37b1-fa89-44de-b8cd-5f6a429347aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icaacc8fc31c18747f58e45b84d7a8a769f2bf012
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com