Revert "Reland "verifier: Cache types by type index.""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b7c51623-1e7e-4a00-abd2-dfefe1fef0ad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61b359a52d06ef4e2ffb8b26c758ea87edfd8882
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com