Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1e4580fa-3ebe-4a4d-8166-e358e25798a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I70daa70e250f0e20b1d0ea639f62de1595aad65b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com