Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dce79378-2b5d-40c3-9390-09253fd7977b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib7cfd35fc31d203554620cca8ad203888f284530
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com