Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b71fdd8e-7d6a-4b64-ac94-bf51712b44f1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I60817360d3241f21169a462853168bb77216d3b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com