Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/02fa0268-c36d-47b6-9487-140d6a51c45c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I79016ca554fd45bda26b825481c7639456c47b7f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com