Revert "Remove interpret_one_instruction from switch interpreter"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/21d330c8-e2fc-482c-b2ac-619e2a94e744). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b70796304a20f2a9c6d8f222a870931d75a822c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com