Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71adbcb8-ad7b-4452-8c6a-0baf4875eef2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9f11744658f8f1131388f9aee69f90aa53649d7d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com