Revert "Avoid move to same registers."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/65ad6530-1a45-4700-aeed-126b19a18df3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f6ccb2ad68e72267f82ab0338b819f9ae732432
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com