Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/10ee720f-57f5-466d-b3d7-1e438b515b4b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0b85e2d6946e4aab99e550d52102dd88aeca8007
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com