Revert "Run RTP after GVN to remove more NullCheck instructions"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92163926-c10d-4722-acf6-50a561262240). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib28cec1902ea8129d33ff29247587fb963b72aa5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com