Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6298ff9-3e0c-43ca-9019-e41b8677d27b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8cf69e50766244d6663bdb9f165418ae4aa715f9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com