Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e33a933-6d07-4133-907d-4452b0f19f9f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4808099aa962d54286b550b9336bbe1c5e724da6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com