Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba389daf-6bcc-478b-a660-9a658ac63542). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic87a96d6df452b147c8681fc1d0ac3eb6000e915
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com