Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/76d0d256-dff9-4291-91cf-a42d57e1549d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I751fe1abc38233b873332e808579e99eaa345979
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com