Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/52b3d455-3fec-410b-b6db-f79ebe0cf12b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5fbd861cc69cb1a17dc2fff10dab0f2c7c556f89
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com