Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7422399c-6bb4-4de0-adef-4a62a80eed53). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3fa3a21f70938354c0918edf15a0b62545268c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com