Revert "Reland^3 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6142a95-5393-4429-8b6c-920dd2e7802f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c7a3293ae7596d39e439d24e762ed01320cca24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com