Revert "run-test: Replace "yes"/"no" with True/False"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/23857cec-4020-40af-93b2-9f980f128242). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25408c44ac2565af9a241d1ee030913f48eb4a3c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com