Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd9bcc77-6204-4699-94a0-00749646d684). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f9813fcb2cd992160f94f8f5a13c1f5fa776b3a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com