Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14d40619-032e-4374-8a8e-f64d6706f840). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I064a277270a4e95ac0e0790b28ca26eb696f3d75
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com