Revert "Do not defer library unloading from zygote"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9d5472e-46e3-4a17-b94e-527dc03ae7be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If95e392ac9c4cdd9c0b33f7ddbed60a0c1e93042
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com