Revert "Reland^3 "Introduce black-dense region in moving space""
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9d5472e-46e3-4a17-b94e-527dc03ae7be). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8bafefd10c56ed3731cb1972f1a482610a2ecc06
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com