Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a9435c6-4972-428d-af86-42f31bf70569). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie1b640b17762db772ff8ed70e19daa4268c04fc0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com