Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a2888ff-97bd-42d6-95d8-87e2b0f5f2a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8f23a61a2fd78c731d263e3c256d1fb0b6a85fc2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com