Revert "Poll for boot image mapping in a heap task."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6a2888ff-97bd-42d6-95d8-87e2b0f5f2a9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36ce031b132fa589632e45bc4c2299fa4ee0d153
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com