Revert "Split dirty-image-objects file between art and framework"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5db79483-4338-49d9-8334-cd0dda534bc4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib2cd459fa65f7751162fbcffc318b7c32169ccba
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com