Revert "Always use an array in the DexCache for ArtField and Art..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/16ee2116-3ff1-42c5-a460-990e08c8ce93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iddf429e6676f14f6325bf975ae5af516d88df5e3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com